myThermistor according to liskov substitution.
