+++ .vscode/c_cpp_properties.json
@@ -1,25 +1,24 @@
 {
-    "configurations": [
-        {
-            "name": "Linux",
-            "includePath": [
-                "/home/dylan/arduino-1.8.13/tools/**",
-                "/home/dylan/Arduino/libraries/**",
-                "/home/dylan/arduino-1.8.13/libraries/**",
-                "/home/dylan/arduino-1.8.13/hardware/tools/**",
-                "/home/dylan/arduino-1.8.13/hardware/arduino/avr/**"
-            ],
-            "forcedInclude": [
-                "/home/dylan/arduino-1.8.13/hardware/arduino/avr/cores/arduino/Arduino.h"
-            ],
-            "defines": [
-                "USBCON"
-            ],
-            "intelliSenseMode": "gcc-x64",
-            "compilerPath": "/usr/bin/gcc",
-            "cStandard": "gnu11",
-            "cppStandard": "gnu++14"
-        }
-    ],
-    "version": 4
+  "configurations": [
+    {
+      "name": "Linux",
+      "includePath": [
+        "/home/dylan/arduino-1.8.13/tools/**",
+        "/home/dylan/Arduino/libraries/**",
+        "/home/dylan/Arduino/libraries/NTC_Thermistor/src/**",
+        "/home/dylan/arduino-1.8.13/libraries/**",
+        "/home/dylan/arduino-1.8.13/hardware/tools/**",
+        "/home/dylan/arduino-1.8.13/hardware/arduino/avr/**"
+      ],
+      "forcedInclude": [
+        "/home/dylan/arduino-1.8.13/hardware/arduino/avr/cores/arduino/Arduino.h"
+      ],
+      "defines": ["USBCON"],
+      "intelliSenseMode": "gcc-x64",
+      "compilerPath": "/usr/bin/gcc",
+      "cStandard": "gnu11",
+      "cppStandard": "gnu++14"
+    }
+  ],
+  "version": 4
 }
