+++ Heater.cpp
@@ -2,7 +2,7 @@
 #include "Flasher.h"
 
 #include "TempSensor.h"
-#include <arduino.h>
+#include <Arduino.h>
 //FLASHER SWITCH CASE STATES
 //==================
 #define IDLE      1
